+++ apps/sim/lib/mcp/oauth/probe.ts
@@ -1,8 +1,8 @@
 import { extractWWWAuthenticateParams } from '@modelcontextprotocol/sdk/client/auth.js'
 import type { FetchLike } from '@modelcontextprotocol/sdk/shared/transport.js'
 import { createLogger } from '@sim/logger'
+import { isLoopbackHostname } from '@sim/security/hostnames'
 import { createPinnedFetch } from '@/lib/core/security/input-validation.server'
-import { isLoopbackHostname } from '@/lib/core/utils/urls'
 import { createSsrfGuardedMcpFetch } from '@/lib/mcp/pinned-fetch'
 import type { McpAuthType } from '@/lib/mcp/types'
